pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "commitguard-cli"
version = "0.2.0"
description = "AI-powered tool to analyze Git commits for bugs and issues"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    {name = "CommitGuard"}
]
keywords = ["git", "commit", "ai", "code-review", "openrouter", "openai", "claude", "gemini", "cli", "commitguard", "commitguard-cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.1.0",
    "gitpython>=3.1.0",
    "openai>=1.0.0",
    "packaging>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/PierrunoYT/commitguard"
Repository = "https://github.com/PierrunoYT/commitguard"
Issues = "https://github.com/PierrunoYT/commitguard/issues"

[project.scripts]
commitguard = "commitguard.cli:main"

[tool.setuptools.packages.find]
include = ["commitguard*"]
